requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
spotipy==2.23.0
python-dotenv==1.0.0 